move hash functions
