+++ native/spark-expr/src/hash_funcs/mod.rs
@@ -1,0 +1,8 @@
+mod sha2;
+pub mod murmur3;
+mod xxhash64;
+pub(super) mod utils;
+
+pub use sha2::{spark_sha224, spark_sha256, spark_sha384, spark_sha512};
+pub use murmur3::spark_murmur3_hash;
+pub use xxhash64::spark_xxhash64;
